+++ examples/server.js
@@ -84,6 +84,9 @@ app.get('/:file', (req, res, file) => {
 // http://127.0.0.1:8080/examples/hello-world/en/js.htm
 // http://127.0.0.1:8080/examples/hello-world/en/web.htm
 // http://127.0.0.1:8080/examples/hello-world/en/web-url-router.htm
+// http://127.0.0.1:8080/examples/hello-world/en/react.htm
+// http://127.0.0.1:8080/examples/hello-world/en/preact.htm
+// http://127.0.0.1:8080/examples/hello-world/en/rax.htm
 app.get('/examples/hello-world/:lang/:file', async (req, res, lang, file) => {
     // CSS or SVG file
     if (!file.endsWith('.htm')) {